Python Services/pyproject.toml
[project]
name = "web-book-store"
version = "0.1.0"
description = "This project involves the development of a virtual book store, implemented using Object Oriented Programming principles and design patterns"
authors = [
    {name = "Cgamez28",email = "cristiangameznunez@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
